Even more cleanup
