+++ shrimpbot.py
@@ -76,8 +76,6 @@ def searchFor(search_term, search_set, match_threshold=100):
     matches = sorted(
         [r for r in ratios], key=lambda ratio: ratio[1] + ratio[2], reverse=True
     )
-    #    if ((int(matches[0][1] + matches[0][2])) > match_threshold) or (int(matches[0][0]) == 100):
-    # logging.info(str(matches[0][1]),str(matches[0][2]))
     if ((int(matches[0][1] + matches[0][2])) > match_threshold) or (
         int(matches[0][1]) == 100
     ):
@@ -148,33 +146,21 @@ async def toggle(ctx):
 async def on_ready():
     BOT_OWNER = bot.get_user(BOT_OWNER_ID)
 
-    logging.info("Logged in as")
-    logging.info(bot.user.name)
-    logging.info(bot.user.id)
-    logging.info("------")
-    logging.info("Owner is")
-    logging.info(BOT_OWNER.name)
-    logging.info(BOT_OWNER.id)
-    logging.info("------")
-    logging.info("Servers using Shrimpbot")
+    logging.info(f"Logged in as{bot.user.name} ({bot.user.id})")
+    logging.info(f"Owner is {BOT_OWNER.name} ({BOT_OWNER.id})")
     for guild in bot.guilds:
-        print("Checking guild {}".format(str(guild)))
-        logging.info(" {}".format(str(guild)))
-        logging.info(" - ID: {}".format(str(guild.id)))
+        logging.info("Server: {} ({})".format(str(guild), str(guild.id)))
         if guild.id == 697833083201650689:
-            print("leaving that one guild")
+            logging.info(f"Leaving {str(guild)}...")
             await guild.leave()
-            logging.info(" [!] LEFT {}".format(str(guild)))
-        if guild.id != 669698762402299904:  # Steel Strat Server are special snowflakes
-            print("Fixing nick in {}".format(str(guild)))
+            logging.info("[!] LEFT {}".format(str(guild)))
+        if guild.id != 669698762402299904:  # BIG Server are special snowflakes
+            logging.info("Fixing my name in {}".format(str(guild)))
             await guild.me.edit(nick="Shrimpbot")
         time.sleep(0.25)
-        print("\n")
-    logging.info("======")
 
+    logging.info("Shrimpbot is online.")
     await bot.change_presence(status=discord.Status.online, activity=note)
-
-    # ~ await bot.edit_profile(username="ShrimpBot")
 
 
 @bot.command()
